Update readme
This changes Nuget.Config to nuget.config in the readme,
corresponding to the recent rename for portability.

The readme still does not document this portability consideration.
It probably should, but I don't have good references or a
sufficiently compelling demonstration of it, yet.

I've also changed some HTML character entities to their literal
Unicode characters, which makes the readme easier to spell-check
and easier to read when reading the raw Markdown code.