Use portable paths in readme
Before, paths were discussed with "/" as the directory separator,
but commands showed "\". My reasoning was that, when I first wrote
it, I was targeting only Windows, so "\" paths would be fine even
outside of PowerShell, and furthermore, PowerShell on Windows
completes paths by writing "\".

However, the inconsistency where "/", as presented, would be
auto-completed to "\" -- even on Windows -- is less severe than the
inconsistency of having the readme use both styles in different
places. I think that has potential to be confusing, and should be
addressed even before dealing with (other blockers for) issue #13.

This makes sense as part of PR #14 because the added PowerShell
script uses "/" everywhere (never "\"). In light of that, keeping
the suggested commands in README.md using "\" would be even more
confusing.

This moves *slightly* toward fixing #13, but the instructions need
to be modified more deeply in order to address Unix-like systems.