Make `permissions` key top-level when there is only one job
For easier reading.

If `permissions` were specified in a workflow with multiple jobs,
then whether it should be top-level, or set individually in the
jobs, or a combination of the two, would depend on whether the jobs
need different permissions. In contrast, in the workflows edited
here, it is just a matter of style.