Go back to ubuntu-22.04 in CI test jobs
This removes the new step to attempt to install the `nuget` CLI on
the `ubuntu-latest` (`ubuntu-24.04`) runner, because it is not in
the APT repositories for Ubuntu 24.04 LTS. It instead changes
`ubuntu-latest` to `ubuntu-24.04` in the `tests` workflow. This
change is done to both the jobs, so that their effects can be
compared, even though `nuget` is only being used for repackaging.