Bring back `global.json` for Infer#
Because it is still needed for `--output` to work.

But modify it to be for .NET 8 rather than .NET 6 as before.

I'm not sure `--output` really is working well enough even with
this for it to be worthwhile, because this does still give:

> The "--output" option isn't supported when building a solution.
> Specifying a solution-level output path results in all projects
> copying outputs to the same directory, which can lead to
> inconsistent builds.

But for now I may keep this, so let the .NET 8 migration changes,
CI changes to make the `nuget` CLI available again, and other
maintenance come through, rather than blocking #62 on `--output`.